+++ lib/routemaster/drain.rb
@@ -1,5 +1,5 @@
 module Routemaster
   module Drain
-    VERSION = '2.4.4'
+    VERSION = '2.5.0'
   end
 end